refactor: Optimize imports and move utils to core
