+++ src/core/utils.ts
@@ -1,4 +1,4 @@
-import { subversionRandom } from "@/utils/base";
+import { hashNumberFromString, seededRandom, subversionRandom } from "@/utils/base";
 import { brandRandom } from "@/utils/equipment";
 import { debounce } from "@/utils/timer";
 
@@ -48,6 +48,38 @@ const notifyContent = debounce(() => {
 
 
 // 
+// --- random ---
+// 
+
+/**
+ * 随机canvas噪音
+ */
+export const randomCanvasNoise = (seed: number) => {
+  const noise: number[] = []
+  for (let i = 0; i < 10; i++) {
+    noise.push(Math.floor(seededRandom(seed++, 255, 0)))
+  }
+  return noise
+}
+
+/**
+ * 获取[x, y]，区间[-1, 1]
+ */
+export const randomWebglNoise = (seed: number): [number, number] => {
+  return [seededRandom(seed, 1, -1), seededRandom(seed + 1, 1, -1)]
+}
+
+/**
+ * 获取随机字体噪音
+ */
+export const randomFontNoise = (seed: number, mark: string): number => {
+  const random = seededRandom((seed + hashNumberFromString(mark)) % Number.MAX_SAFE_INTEGER, 3, 0)
+  if ((random * 10) % 1 < 0.9) return 0;
+  return Math.floor(random) - 1;
+}
+
+
+// 
 // --- other ---
 // 
 
@@ -78,11 +110,12 @@ export const drawNoise = (
   let noiseIndex = 0;
   let isChanged = false
 
-  const center = new Uint8ClampedArray(4)
-  const up = new Uint8ClampedArray(4)
-  const down = new Uint8ClampedArray(4)
-  const left = new Uint8ClampedArray(4)
-  const right = new Uint8ClampedArray(4)
+  const Arr = Uint8ClampedArray;
+  const center = new Arr(4)
+  const up = new Arr(4)
+  const down = new Arr(4)
+  const left = new Arr(4)
+  const right = new Arr(4)
 
   const pixelData = imageData.data
